Move transcript widgets into fresh-agent
