+++ src/components/fresh-agent/FreshAgentItemCard.tsx
@@ -1,7 +1,7 @@
 import { useMemo, useState } from 'react'
 import { Check, ChevronRight, Loader2, X } from 'lucide-react'
-import DiffView from '@/components/agent-chat/DiffView'
-import { getToolPreview } from '@/components/agent-chat/tool-preview'
+import DiffView from '@/components/fresh-agent/shared/DiffView'
+import { getToolPreview } from '@/components/fresh-agent/shared/tool-preview'
 import { LazyMarkdown } from '@/components/markdown/LazyMarkdown'
 import type { FreshAgentTranscriptItem } from '@shared/fresh-agent-contract'
 import { cn } from '@/lib/utils'
