+++ src/components/fresh-agent/shared/DiffView.tsx
