+++ src/components/fresh-agent/shared/SlotReel.tsx
