+++ src/components/fresh-agent/shared/tool-preview.ts
