+++ test/unit/client/components/agent-chat/DiffView.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, afterEach } from 'vitest'
 import { render, screen, cleanup } from '@testing-library/react'
-import DiffView from '../../../../../src/components/agent-chat/DiffView'
+import DiffView from '@/components/fresh-agent/shared/DiffView'
 
 describe('DiffView', () => {
   afterEach(cleanup)
