+++ test/unit/client/components/agent-chat/tool-preview.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest'
-import { getToolPreview } from '@/components/agent-chat/tool-preview'
+import { getToolPreview } from '@/components/fresh-agent/shared/tool-preview'
 
 describe('getToolPreview', () => {
   it('returns empty string when no input', () => {
